[ADD] readsize익스텐션추가(#11)
