+++ Genti/Genti/Extentions/View+Ext.swift
@@ -1,0 +1,31 @@
+//
+//  View+Ext.swift
+//  Genti
+//
+//  Created by uiskim on 4/29/24.
+//
+
+import SwiftUI
+
+extension View {
+  @ViewBuilder
+  func onReadSize(_ perform: @escaping (CGSize) -> Void) -> some View {
+    self.customBackground {
+      GeometryReader { geometryProxy in
+        Color.clear
+          .preference(key: SizePreferenceKey.self, value: geometryProxy.size)
+      }
+    }
+    .onPreferenceChange(SizePreferenceKey.self, perform: perform)
+  }
+  
+  @ViewBuilder
+  func customBackground<V: View>(alignment: Alignment = .center, @ViewBuilder content: () -> V) -> some View {
+    self.background(alignment: alignment, content: content)
+  }
+}
+
+struct SizePreferenceKey: PreferenceKey {
+  static var defaultValue: CGSize = .zero
+  static func reduce(value: inout CGSize, nextValue: () -> CGSize) { }
+}
